processing: move the datecode-range helper into utils/io.py, drop _common.py
_common.py held a single tools-free helper, _datecode_dirs, used only by the
masters orchestrator to expand a --date_range against the L0 tree. It is a
filesystem-discovery helper that belongs beside load_junk_obs_ids and the
product-path builders in kpfpipe.utils.io, not in a scripts-local module.

Rename it datecode_dirs_in_range and move it into kpfpipe/utils/io.py (a
downward import into the shared kpfpipe layer, which scripts/ may depend on);
masters.py now imports it from there. Its test moves from test_common.py into
TestDatecodeDirsInRange in test_io.py. Delete _common.py and test_common.py.
Docs (CLAUDE.md CLI-architecture note, test_masters_script.py header) updated
to point at the new home.

rv_timeseries.py keeps its own local _datecode_dirs copy and is unaffected
(it remains the deliberately off-limits, runtime-broken driver).
